+++ src/main/java/kitchenpos/ui/request/MenuProductRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.ui.dto;
+package kitchenpos.ui.request;
 
 public class MenuProductRequest {
     private Long productId;
