+++ src/main/java/kitchenpos/ui/request/OrderLineItemRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.ui.dto;
+package kitchenpos.ui.request;
 
 public class OrderLineItemRequest {
     private Long menuId;
