+++ src/main/java/kitchenpos/ui/request/OrderRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.ui.dto;
+package kitchenpos.ui.request;
 
 import java.util.List;
