+++ src/main/java/kitchenpos/ui/request/OrderTableIdRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.ui.dto;
+package kitchenpos.ui.request;
 
 public class OrderTableIdRequest {
     private Long id;
