+++ src/main/java/kitchenpos/ui/request/OrderTableNumberOfGuestRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.ui.dto;
+package kitchenpos.ui.request;
 
 public class OrderTableNumberOfGuestRequest {
     private Integer numberOfGuests;
