+++ src/main/java/kitchenpos/ui/request/OrderTableRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.ui.dto;
+package kitchenpos.ui.request;
 
 public class OrderTableRequest {
     private int numberOfGuests;
